chore: refactor oc-table-body component using script setup
