+++ packages/design-system/src/components/OcTable/OcTable.vue
@@ -98,7 +98,7 @@
 </template>
 <script lang="ts" setup>
 import OcThead from '../_OcTableHeader/_OcTableHeader.vue'
-import OcTbody from '../_OcTableBody/_OcTableBody.vue'
+import OcTbody from '../OcTableBody/OcTableBody.vue'
 import OcTr from '../_OcTableRow/_OcTableRow.vue'
 import OcTh from '../_OcTableCellHead/_OcTableCellHead.vue'
 import OcTd from '../_OcTableCellData/_OcTableCellData.vue'
