+++ packages/design-system/src/components/OcTableBody/OcTableBody.vue
@@ -3,9 +3,8 @@
     <slot />
   </tbody>
 </template>
-<script lang="ts">
-import { defineComponent } from 'vue'
-export default defineComponent({
+<script lang="ts" setup>
+defineOptions({
   name: 'OcTbody',
   status: 'ready',
   release: '2.1.0'
