+++ packages/design-system/src/components/index.ts
@@ -39,7 +39,7 @@ export { default as OcSpinner } from './OcSpinner/OcSpinner.vue'
 export { default as OcStatusIndicators } from './OcStatusIndicators/OcStatusIndicators.vue'
 export { default as OcSwitch } from './OcSwitch/OcSwitch.vue'
 export { default as OcTable } from './OcTable/OcTable.vue'
-export { default as OcTableBody } from './_OcTableBody/_OcTableBody.vue'
+export { default as OcTableBody } from './OcTableBody/OcTableBody.vue'
 export { default as OcTableCell } from './_OcTableCell/_OcTableCell.vue'
 export { default as OcTableCellData } from './_OcTableCellData/_OcTableCellData.vue'
 export { default as OcTableCellHead } from './_OcTableCellHead/_OcTableCellHead.vue'
